fix litter sound and resolve "0x08fc1000" bug...
git-svn-id: http://xreader.googlecode.com/svn/branches/trunk-alpha-me@1493 9e70b50a-5c3c-0410-8115-a7454348afc8

+++ src/scene_text.c
@@ -1629,7 +1629,7 @@ bool scene_bookmark(PBookViewData pView)
 				  menubcolor, true, scene_bookmark_predraw, NULL,
 				  scene_bookmark_menucb)) != INVALID);
 
-	return (bool) item[1].data;
+	return (bool)(int) item[1].data;
 }
 
 int book_handle_input(PBookViewData pView, dword * selidx, dword key)
